requirements.txt
tqdm
langsmith
chromadb
langchain
langchain-community
langchain-chroma
langchain-ollama
pypdf
gardio